+++ engine/src/main/kotlin/engine/expressions/ExpressionCreators.kt
@@ -77,6 +77,18 @@ fun simplifiedNegOf(expr: Expression) = when (expr) {
     else -> negOf(expr)
 }
 
+/**
+ * The simplified negation of an expression, but if [expr] is a sum then each of its terms is negated.  So
+ *
+ *     x - 2 --> -x + 2
+ *
+ * Note: it remains to be determined what a good origin would be for the terms in the result.
+ */
+fun simplifiedNegOfSum(expr: Expression) = when (expr) {
+    is Sum -> sumOf(expr.children.map { simplifiedNegOf(it).withOrigin(Introduce(listOf(it))) })
+    else -> simplifiedNegOf(expr).withOrigin(Introduce(listOf(expr)))
+}
+
 fun plusOf(expr: Expression) = buildExpression(UnaryExpressionOperator.Plus, listOf(expr))
 
 fun plusMinusOf(expr: Expression) = buildExpression(UnaryExpressionOperator.PlusMinus, listOf(expr))